+++ packages/utils/src/index.ts
@@ -1,7 +1,6 @@
 import amount from './amount';
 import cachedThrottle from './cached-throttle';
 import crypto from './crypto';
-import currency from './currency';
 import encryption from './encryption';
 import identity from './identity';
 import retry from './retry';
@@ -16,7 +15,6 @@ export default {
   amount,
   cachedThrottle,
   crypto,
-  currency,
   deepCopy: utils.deepCopy,
   deepSort: utils.deepSort,
   encryption,