+++ packages/payment-detection/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.30.2](https://github.com/RequestNetwork/requestNetwork/compare/@requestnetwork/payment-detection@0.30.0...@requestnetwork/payment-detection@0.30.2) (2020-12-22)
+
+**Note:** Version bump only for package @requestnetwork/payment-detection
+
+
+
+
+
+## [0.30.1](https://github.com/RequestNetwork/requestNetwork/compare/@requestnetwork/payment-detection@0.30.0...@requestnetwork/payment-detection@0.30.1) (2020-12-21)
+
+**Note:** Version bump only for package @requestnetwork/payment-detection
+
+
+
+
+
 # 0.30.0 (2020-12-02)