+++ packages/transaction-manager/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.26.2](https://github.com/RequestNetwork/requestNetwork/compare/@requestnetwork/transaction-manager@0.26.0...@requestnetwork/transaction-manager@0.26.2) (2020-12-22)
+
+**Note:** Version bump only for package @requestnetwork/transaction-manager
+
+
+
+
+
+## [0.26.1](https://github.com/RequestNetwork/requestNetwork/compare/@requestnetwork/transaction-manager@0.26.0...@requestnetwork/transaction-manager@0.26.1) (2020-12-21)
+
+**Note:** Version bump only for package @requestnetwork/transaction-manager
+
+
+
+
+
 # [0.26.0](https://github.com/RequestNetwork/requestNetwork/compare/@requestnetwork/transaction-manager@0.8.1...@requestnetwork/transaction-manager@0.26.0) (2020-12-02)