+++ packages/web3-signature/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.28](https://github.com/RequestNetwork/requestNetwork/compare/@requestnetwork/web3-signature@0.4.26...@requestnetwork/web3-signature@0.4.28) (2020-12-22)
+
+**Note:** Version bump only for package @requestnetwork/web3-signature
+
+
+
+
+
+## [0.4.27](https://github.com/RequestNetwork/requestNetwork/compare/@requestnetwork/web3-signature@0.4.26...@requestnetwork/web3-signature@0.4.27) (2020-12-21)
+
+**Note:** Version bump only for package @requestnetwork/web3-signature
+
+
+
+
+
 ## [0.4.26](https://github.com/RequestNetwork/requestNetwork/compare/@requestnetwork/web3-signature@0.4.5...@requestnetwork/web3-signature@0.4.26) (2020-12-02)